+++ Tugas4/src/Dot.h
@@ -9,22 +9,22 @@ using namespace std;
 class Dot {
   public:
     Dot();
-    Dot(int, int, Color);
+    Dot(float, float, Color);
     //Dot(int, int, string);
     //Dot(const Dot&);
     //~Dot();
     //Dot& operator=(const Dot&);
-    void setCoordinate(int,int);
-    void setX(int);
-	  int getX();
-	  void setY(int);
-    int getY();
+    void setCoordinate(float,float);
+    void setX(float);
+	  float getX();
+	  void setY(float);
+    float getY();
     void setColor(int r, int g, int b);
 	  Color getColor();
 
   //protected:
-    int x;
-    int y;
+    float x;
+    float y;
     Color color;
 };
